Funkcijų ir struktūrų aprašai sukelti į header failus.
