+++ skaiciavimai.h
@@ -1,0 +1,10 @@
+#ifndef SKAICIAVIMAI_H
+#define SKAICIAVIMAI_H
+
+#include <vector>
+
+float vidurkis(std::vector<int> pazymiai);
+float mediana(std::vector<int> pazymiai);
+float galutinis(float paz_agg, int egz_paz);
+
+#endif
